added e2e test for operator resolution, fixed unit-test gh action running e2e tests
